Update ThreadExample to 2021.0 project template
